+++ llvm/unittests/TargetParser/RISCVISAInfoTest.cpp
@@ -1110,6 +1110,8 @@ Experimental extensions
     zalasr               0.1
     zvbc32e              0.7
     zvkgs                0.7
+    sdext                1.0
+    sdtrig               1.0
     smctr                1.0
     ssctr                1.0
     svukte               0.3